Demo/service-examples/stress-ng/requirements.txt
pulsar-client~=2.7.2